Permission Checks/ Various Fixes

+++ commands/echo.js
@@ -1,6 +1,14 @@
 exports.run = async (client, message, args, level) => {
-	message.delete();
-	message.channel.send(args.join(" "));
+
+	if(message.channel.memberPermissions(message.guild.me).has("MANAGE_MESSAGES")){
+		message.delete();
+	}
+
+	const joinargs = args.join(" ");
+
+	if(!joinargs) return message.channel.send("Usage: [echo](<..text>)", { code: "markdown" });
+
+	message.channel.send(joinargs);
 };
 
 exports.conf = {
@@ -14,5 +22,5 @@ exports.help = {
 	name: "echo",
 	category: "Misc",
 	description: "Repeats what you say",
-	usage: "echo](<..variable>)",
+	usage: "echo](<..text>)",
 };
